Implement last better
Yes, that means the actually accepted parameters to the callables are moved further away fro its documented limitations.
But there's a precedent, in my view, in the fact that Always is replaced by a Displayable in LayeredImage.add
So I think this is acceptable.
And touching dicts was weird and cost more time than required.